(GH-985) Addition of Add to Script Builder Feature
A new feature "Add to Script Builder" has been added to the packages
area. A "plus" sign now appears next to each package including each
package version. A user will now be able to select all the packages
that he/she needs and they will be added to a list located on the left
side of the page. From inside the list the user can delete packages no
longer needed, export the list as a packages.config file, or have a
script generated based on the packages in their builder.