(layout) Resource Filtering
Allows resources within the resource area to be filtered by tag or
title. A search box has been added for users to search for videos based
on these items. The tags for each resource also now appear on their
card on the list screen, and each individual post. The style of this
matches how the package area is handling tags.