(maint) Refactor testimonials
The testimonials that were added to the resource area in a recent PR
were put in as a long list, but now they will be handled just like any
of the other resources (case studies, videos, success stories). Each
testimonial md file will start with a `04` to identify that it is a
testimonial.