(layout) Update Smooth Scrolling JS
Updates the smooth scroll JS to ensure that if a page has a sticky
navigation on the top of the page, that the height of that nav bar is
considered when scrolling to an on page link. Without this, the scroll
position will always be off by the height of the sticky nav and cover
up some of the content. If there is no sticky nav on the page, the
scrolling works as it always has.